requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
email-validator
pytest
httpx
